Finished first attempt at multiple tools with different axis motors.
Removed some old code that was used for changing the motor select pin.
Not sure how "class" vs "include" works in Kernal, but it compiles so...
Compiled using:
make all AXIS=4 CNC=1

Might need to look in to excluding more modules. Who needs a screen
anyway?
Might create some sort of array in the future for handling the changing
of multiple axis motors.